[LCR-341] refactor: NotificationIntegrationTest 위치 수정
